refactor(core): move the crate-root, metrics/, gpu/, storage/, update_check/, agent/ and simd_native/ inline test modules to sibling files (#1918)
Fourth and last lot for velesdb-core. Twenty-five files; the two shapes
the audit had flagged as the hardest are both here.

test_fixtures.rs holds a pub(crate) mod fixtures, not a private mod
tests, and eight test files across the crate import
crate::test_fixtures::fixtures::{setup_collection, make_point_with_payload,
...}. The extraction tool refuses this shape, so it was moved by hand and
the declaration keeps pub(crate) verbatim; dropping the visibility would
have broken all eight.

simd_neon.rs carries its block under #[cfg(all(test, target_arch =
"aarch64"))], which an x86 build never compiles, so the test-name diff
says nothing about it. Rather than ship it unverified, the aarch64 std
component was installed and the whole crate cross-checked with
cargo check --target aarch64-unknown-linux-gnu --profile test, which does
compile the block behind that gate.

Five files already had a <stem>_tests.rs sibling (config, distance,
gpu_csr, half_precision, simd_dispatch); config.rs and distance.rs name
their module shared_toml_tests / distance_metric_names_tests, so those
names resolve the collision on their own and the other three take
_unit_tests.rs.

The sorted test-name list is byte-identical before and after, re-checked
after rustfmt: 5357 tests, 5341 passing and 16 pre-existing ignored.
