+++ scripts/inline-tests-baseline.txt
@@ -3,34 +3,9 @@ crates/tauri-plugin-velesdb/src/error.rs	128
 crates/tauri-plugin-velesdb/src/events.rs	37
 crates/tauri-plugin-velesdb/src/helpers.rs	200
 crates/tauri-plugin-velesdb/src/lib.rs	36
-crates/velesdb-core/src/agent/memory_helpers.rs	160
 crates/velesdb-core/src/api_types/requests.rs	65
-crates/velesdb-core/src/config.rs	103
-crates/velesdb-core/src/distance.rs	35
-crates/velesdb-core/src/fault_injection.rs	48
 crates/velesdb-core/src/filter/matching.rs	292
 crates/velesdb-core/src/filter/mod.rs	119
-crates/velesdb-core/src/gpu/gpu_csr.rs	141
-crates/velesdb-core/src/gpu/gpu_traversal.rs	229
-crates/velesdb-core/src/gpu/pq_gpu.rs	119
-crates/velesdb-core/src/half_precision.rs	73
-crates/velesdb-core/src/metrics/guardrails.rs	100
-crates/velesdb-core/src/metrics/latency.rs	38
-crates/velesdb-core/src/metrics/operational.rs	97
-crates/velesdb-core/src/metrics/query.rs	108
-crates/velesdb-core/src/metrics/retrieval.rs	76
-crates/velesdb-core/src/scored_result.rs	57
-crates/velesdb-core/src/simd_dispatch.rs	276
-crates/velesdb-core/src/simd_native/adc.rs	118
-crates/velesdb-core/src/simd_neon.rs	77
-crates/velesdb-core/src/simd_neon_prefetch.rs	53
-crates/velesdb-core/src/storage/async_ops.rs	35
-crates/velesdb-core/src/storage/atomic_write.rs	57
-crates/velesdb-core/src/test_fixtures.rs	51
-crates/velesdb-core/src/update_check/check.rs	26
-crates/velesdb-core/src/update_check/config.rs	132
-crates/velesdb-core/src/update_check/instance_id.rs	46
-crates/velesdb-core/src/validation.rs	101
 crates/velesdb-core/src/velesql/ast/mod.rs	116
 crates/velesdb-core/src/velesql/cache.rs	241
 crates/velesdb-core/src/velesql/cost_estimator/plan_cost.rs	298
